+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change log
 
+## master
+ - Made `data` argument to the table `__init__()` a keyword argument to make inheritance easier. Will raise a `TypeError` if not passed.
+
 ## 1.12.0 (2017-10-10)
  - Allow export filename customization [#484](https://github.com/bradleyayers/django-tables2/pull/484) by [@federicobond](https://github.com/federicobond)
  - Fixed a bug where template columns were not rendered for pinned rows ([#483](https://github.com/bradleyayers/django-tables2/pull/483) by [@khirstinova](https://github.com/khirstinova), fixes [#482](https://github.com/bradleyayers/django-tables2/issues/482))